🚧 M5G987 task: apply external agent result
