+++ .agentplane/tasks/202609042338-M5G987/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202609042338-M5G987/repair-atomic-scope-extension-projection-and-acc",
   "created_at": "2026-09-05T00:35:18.474Z",
-  "diffstat_sha256": "sha256:026ec2712d8575e61a72a088984b99ecbc5d2e544960051493ce0678b5f2cfe4",
+  "diffstat_sha256": "sha256:b851503d3edd952305caed9fe311189c63117cf22a2685832522fa54e3544255",
   "last_verified_at": null,
   "provider": {
     "hostname": "github.com",
