+++ packages/agentplane/src/commands/shared/workflow-step-factory.ts
@@ -1,4 +1,5 @@
 import type { TaskData } from "../../backends/task-backend.js";
+import { hasUninitializedTaskBaseline } from "./workflow-step-policy-scope.js";
 import { isRecord } from "../../shared/guards.js";
 import type { RouteBlocker } from "./route-oracle.js";
 import {
@@ -319,7 +320,10 @@ export function directStep(state: WorkflowRouteState): WorkflowStep {
       selectedBlocker: routeBlockerFor(state, "runner_alive"),
     });
   }
-  if (String(state.task.status).toUpperCase() !== "DOING") {
+  if (
+    String(state.task.status).toUpperCase() !== "DOING" ||
+    hasUninitializedTaskBaseline(state.task)
+  ) {
     const body = "Start: continue direct-mode task in current checkout.";
     return cliOperationStep({
       state,
